logistics_project/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15,<4.0
djangorestframework-simplejwt>=5.3,<6.0
django-cors-headers>=4.3,<5.0
pandas>=2.2,<3.0
openpyxl>=3.1,<4.0
reportlab>=4.0,<5.0
Pillow>=10.0,<12.0
google-generativeai>=0.8,<1.0
python-dotenv>=1.0,<2.0
